Refactored implemntation of try_from() on the Point Struct
Rewritten the code so that it takes a string slice from the second
character to the end of the string instead of just the second character.
Previously it couldn't handle a10 as a coordinate. Also implemented
bounds so that it returns an error if it is greater than the board size.
Made clippy happy too