The game grid is now drawn in place
The game grid is now drawn in place by jumping up however many lines that
were printed. Opted to do it this way as when I tried saving the cursor
position and restoring it, it would sometimes redraw with an 'afterimage'
I reckon this is because the SavePosition function in Crossterm is
relative to the terminals current window, rather than the actual line
number of the terminal. There might be an easier way to do this, however
for now it works just fine. Also formated main.rs, and deleted
output.txt which shouldn'tve been tracked anyway