Started implementing an algorithm for placing battleships
Currently got a basic algorithm that places the battleships randomly on
the board, with no overlaps and all within the bounds of the board.
Currently however the placement seems... off? I think I need to add a
buffer around each placed ship, so that the ships are not all placed
together in bunches. This commit touches a lot of files because I was a
little stupid and didn't break them up. Made some modules public and
changed main.rs temporarily so I could visualise the placement
algorithm.