Buffer points sometimes are not generated
Added a chance that buffer points will not be generated, which is
controlled via a constant. Not sure if the current value is right, might
require some tweaking as it should be fairly rare. Also noticed that the
ships being generated were 1 segment longer than they should be so
fixed that bug. I don't actually like the shorter ships so I might
change the length of the ships being generated