Added colours to the guesses
Hits are now printed as red, and guesses are now printed as yellow.
However there is a weird issue that when a guess is coloured, the cell
walls 'shrink', making the grid look off. When play testing one round
the 7th row printed off with weird characters pushing it out of
alignment. Could be an issue with the library itself? The issue happens
with gnome terminal, the neovide client, and kitty so I don't think its
an issue with my terminal