Fixed Border Shrinking Error
Fixed the border chars 'shrinking' when an X or a M was printed inside
of them. Turns out it was the format! Macro, however I don't exactly know
why? As I couldn't use the format library I had to use inline escape
sequences to colour the hits and misses. Works really well however, and the
issue is now fixed